feat(wallet): match v2 Figma layout — three tabs + right sidebar, gate web3 behind NEXT_PUBLIC_WEB3
Rewrite /wallet page to match the original static Figma design:
- Three full-width tabs (Balance, P&L, Positions) as the main content
- Right sidebar with search, wallet preview, and latest positions
- Remove on-chain tabs (overview, tokens, NFTs, activity) from default mode
- Add NEXT_PUBLIC_WEB3 env var (default: false) to gate on-chain features
- Extract WalletPreviewWidget into its own file for reuse
