+++ .env.example
@@ -847,6 +847,8 @@ NEXT_PUBLIC_DISABLE_SENTRY=
 NEXT_PUBLIC_STATIC_ASSETS_URL=
 NEXT_PUBLIC_TREASURY_ADDRESS=
 NEXT_PUBLIC_WAITLIST_MODE=
+# Web3 wallet mode — enables on-chain tabs (overview, tokens, NFTs, activity). Default: false
+NEXT_PUBLIC_WEB3=false
 NFT_BASE_URI=
 NFT_COLLECTION_SIZE=
 NFT_CONTRACT_DEPLOY_BLOCK=
